userspace/nat: split NAT64 compiler into nat64.go
Continues the #4421 (codex-173 #5) pure code-motion split. Moves
buildNAT64Snapshots verbatim out of nat.go into nat64.go. This is the
IPv6->IPv4 (natv6v4) snapshot builder, including the global
no-v6-frag-header DF-policy replication onto every emitted rule.

No body edits, no renames. The builder has no type-local helpers beyond
itself and depends only on config, so nat64.go carries a single import.

Validation: go build ./... green.
