+++ pkg/dataplane/userspace/nat64.go
@@ -1,0 +1,48 @@
+package userspace
+
+import (
+	"github.com/psaab/xpf/pkg/config"
+)
+
+func buildNAT64Snapshots(cfg *config.Config) []NAT64RuleSnapshot {
+	if cfg == nil || len(cfg.Security.NAT.NAT64) == 0 {
+		return nil
+	}
+	// `security nat natv6v4 no-v6-frag-header` is a global option, but the
+	// dataplane consumes NAT64 state per rule-set. Replicate the flag onto
+	// every emitted rule so the IPv6->IPv4 translator can honor it. The option
+	// is an option-gated LOCAL DF policy (not the size-driven RFC 7915 5.1
+	// selection): when set the translator clears DF so the IPv4 packet stays
+	// fragmentable (DF=0, non-atomic) and carries a generated non-zero,
+	// non-repeating Identification (RFC 6864 4.1) rather than the default DF=1
+	// atomic framing.
+	noV6FragHeader := cfg.Security.NAT.NATv6v4 != nil && cfg.Security.NAT.NATv6v4.NoV6FragHeader
+	out := make([]NAT64RuleSnapshot, 0, len(cfg.Security.NAT.NAT64))
+	for _, rs := range cfg.Security.NAT.NAT64 {
+		if rs == nil || rs.Prefix == "" {
+			continue
+		}
+		// #2214: initialize non-nil so a rule with no resolvable source pool
+		// marshals `pool_addresses` as `[]`, never JSON `null`. The field has
+		// no `,omitempty` (an empty pool is still a meaningful "no source-pool
+		// resolved" state the dataplane must see), and the Rust `Vec<String>`
+		// rejects an explicit null — which aborts the whole snapshot decode and
+		// kills ALL transit (#1961 no-transit signature).
+		poolAddresses := []string{}
+		if rs.SourcePool != "" {
+			if pool, ok := cfg.Security.NAT.SourcePools[rs.SourcePool]; ok && pool != nil {
+				if pool.Address != "" {
+					poolAddresses = append(poolAddresses, pool.Address)
+				}
+				poolAddresses = append(poolAddresses, pool.Addresses...)
+			}
+		}
+		out = append(out, NAT64RuleSnapshot{
+			Name:           rs.Name,
+			Prefix:         rs.Prefix,
+			PoolAddresses:  poolAddresses,
+			NoV6FragHeader: noV6FragHeader,
+		})
+	}
+	return out
+}
